eLibros/requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
gunicorn
idna
oauthlib
packaging
Pillow
psycopg
psycopg-binary
pycparser
PyJWT
python3-openid
requests
requests-oauthlib
sqlparse
typing_extensions
urllib3
whitenoise
dj-database-url
six
django-simple-history
django-extensions
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-filter-stubs
django-allauth
dj-rest-auth[with_social]